Added Vega Calibration Code and Files
The calibration for pur spectrometer includes a spectrum calibration done using Veg Model Spectrum as sourced from SBPy (based on HST observations). This code uses the model and our processed spectrum to assign weights to the recorded wavelength. This was important to be done since our planetary sesnor was more sensitive to Visible Red and Near IR region of the spectrum.